requirements.txt
python-dotenv
streamlit==1.38.0
pip-system-certs